feat: Create interface folder to store component interfaces Removed interface code from Button.tsx and Input-field.tsx and placed them in their corresponding .ts files
