+++ src/components/interfaces/InputFieldProps.ts
@@ -1,0 +1,10 @@
+export interface Field {
+    id: string;
+    label: string;
+    variant: 'outlined' | 'filled' | 'standard';
+    defaultValue?: string;
+}
+
+export interface Props {
+    fields: Field[];
+}
